Rename user event views
