--- src/views/admin/Report/ReportsView.vue
@@ -1,10 +1,0 @@
-<script setup>
-import { ref, computed } from 'vue'
-import { useRouter } from 'vue-router'
-import AdminLayout from '@/layouts/AdminLayout.vue'
-</script>
-
-<template>
-    <AdminLayout/>
-
-</template>
