+++ src/views/event/EventDetailsView.vue
